group

Module wnaf

Structs

LimbBuffer 🔒

This struct represents a view of a sequence of bytes as a sequence of u64 limbs in little-endian byte order. It maintains a current index, and allows access to the limb at that index and the one following it. Bytes beyond the end of the original buffer are treated as zero.

Wnaf

A “w-ary non-adjacent form” scalar multiplication (also known as exponentiation) context.

WnafBase

A fixed window table for a group element, precomputed to improve the speed of scalar multiplication.

WnafScalar

A “w-ary non-adjacent form” scalar, that uses precomputation to improve the speed of scalar multiplication.

Traits

WnafGroup

Extension trait on a Group that provides helpers used by Wnaf.

Functions

wnaf_exp 🔒

Performs w-NAF exponentiation with the provided window table and w-NAF form scalar.

wnaf_form 🔒

Replaces the contents of wnaf with the w-NAF representation of a little-endian scalar.

wnaf_table 🔒

Replaces the contents of table with a w-NAF window table for the given window size.