Crate group

Re-exports

pub use ff;

Modules

cofactor

prime

Structs

Wnaf

A “w-ary non-adjacent form” scalar multiplication (also known as exponentiation) context.

WnafBase

A fixed window table for a group element, precomputed to improve the speed of scalar multiplication.

WnafScalar

A “w-ary non-adjacent form” scalar, that uses precomputation to improve the speed of scalar multiplication.

Traits

Curve

Efficient representation of an elliptic curve point guaranteed.

Group

This trait represents an element of a cryptographic group.

GroupEncoding

GroupOps

A helper trait for types with a group operation.

GroupOpsOwned

A helper trait for references with a group operation.

ScalarMul

A helper trait for types implementing group scalar multiplication.

ScalarMulOwned

A helper trait for references implementing group scalar multiplication.

UncompressedEncoding

Affine representation of a point on an elliptic curve that has a defined uncompressed encoding.

WnafGroup

Extension trait on a Group that provides helpers used by Wnaf.