Trait group::WnafGroup

pub trait WnafGroup: Group {
    // Required method
    fn recommended_wnaf_for_num_scalars(num_scalars: usize) -> usize;
}

Extension trait on a Group that provides helpers used by Wnaf.

Required Methods

fn recommended_wnaf_for_num_scalars(num_scalars: usize) -> usize

Recommends a wNAF window size given the number of scalars you intend to multiply a base by. Always returns a number between 2 and 22, inclusive.

Object Safety

This trait is not object safe.